implement alert position bottom
In doing so, opportunity to correct onMoveShouldSetPanResponder to be based on panResponderEnabled hence eliminates need for panResponderMoveDistance prop. This was based on: "Called for every touch move on the View when it is not the responder: does this view want to "claim" touch responsiveness?"  for this case yes when enabled. https://reactnative.dev/docs/gesture-responder-system#responder-lifecycle